32 Gflops float operation on GTX1080 with ldoor matrix

+++ solver_test.c
@@ -5,7 +5,7 @@
 #include "solver.h"
 #include <unistd.h>
 
-static int matrixRead(matrixCOO* localMatrixCOO, double** xCompare_in, double** y_in, FILE *f)
+static int matrixRead(matrixCOO* localMatrixCOO, float** xCompare_in, float** y_in, FILE *f)
 {
 	int _dimension, _N, _lowerNum;
 	int ret_code, totalNum, lowerNum;
@@ -15,10 +15,10 @@ static int matrixRead(matrixCOO* localMatrixCOO, double** xCompare_in, double** 
 	lowerNum = _lowerNum;
 	totalNum = lowerNum*2-_dimension;
 	
-	*xCompare_in = (double*)malloc(_dimension*sizeof(double));
-	*y_in = (double*)malloc(_dimension*sizeof(double));
-	double*	xCompare = *xCompare_in;
-	double*	y = *y_in;
+	*xCompare_in = (float*)malloc(_dimension*sizeof(float));
+	*y_in = (float*)malloc(_dimension*sizeof(float));
+	float*	xCompare = *xCompare_in;
+	float*	y = *y_in;
 	/*The overall number of nozeros in this matrix*/
 	localMatrixCOO->totalNum = totalNum;
 	localMatrixCOO->nParts = ceil(((float) _dimension)/vectorCacheSize);
@@ -30,22 +30,22 @@ static int matrixRead(matrixCOO* localMatrixCOO, double** xCompare_in, double** 
 
 	int* lowerI=(int *) malloc(lowerNum*sizeof(int));
 	int* lowerJ=(int *) malloc(lowerNum*sizeof(int));
-	double* lowerV=(double *) malloc(lowerNum*sizeof(double));
+	float* lowerV=(float *) malloc(lowerNum*sizeof(float));
 
 	localMatrixCOO->I=(int *) malloc(totalNum*sizeof(int));
 	localMatrixCOO->J=(int *) malloc(totalNum*sizeof(int));
-	localMatrixCOO->V=(double *) malloc(totalNum*sizeof(double));
-	localMatrixCOO->diag = (double *) malloc(_dimension*sizeof(double));
+	localMatrixCOO->V=(float *) malloc(totalNum*sizeof(float));
+	localMatrixCOO->diag = (float *) malloc(_dimension*sizeof(float));
 	int* numInRow = localMatrixCOO->numInRow;
 	int* I = localMatrixCOO->I;
 	int* J = localMatrixCOO->J;
-	double* V = localMatrixCOO->V;
-	double* diag = localMatrixCOO->diag;
+	float* V = localMatrixCOO->V;
+	float* diag = localMatrixCOO->diag;
 	
 	int tempI, tempJ;
-	double tempV;
+	float tempV;
 	for (int i=0; i<lowerNum; i++){
-		fscanf(f, "%d %d %lg\n", &tempI, &tempJ, &tempV);
+		fscanf(f, "%d %d %f\n", &tempI, &tempJ, &tempV);
 		lowerJ[i]=tempJ-1;  /* adjust from 1-based to 0-based */
 		lowerI[i]=tempI-1;
 		lowerV[i]=tempV;
@@ -75,7 +75,7 @@ static int matrixRead(matrixCOO* localMatrixCOO, double** xCompare_in, double** 
 	numInRow[_dimension-1]=0;
 	for (int i=0;i < _dimension;i++){
 		srand(i);
-		xCompare[i]=(double) (rand()%200-100)/100;
+		xCompare[i]=(float) (rand()%200-100)/100;
 		//x_compare[i]=1;
 	}
 	int index1, index2;
@@ -117,9 +117,9 @@ int main(int argc, char* argv[])
 	MM_typecode matcode;
 	int MAXIter = 0;
 	FILE *f;
-	double *x;
-	double *y;
-	double *xCompare;
+	float *x;
+	float *y;
+	float *xCompare;
 	char fileName[100];
 	fileName[0] = '\0';
 	int oc;
@@ -195,8 +195,8 @@ int main(int argc, char* argv[])
 	matrixCOO localMatrixCOO;
 	matrixRead(&localMatrixCOO, &xCompare, &y, f);
 	fclose(f);
-	double *xReorder = (double* )calloc(localMatrixCOO.dimension, sizeof(double)); 
-	double *yReorder = (double* )calloc(localMatrixCOO.dimension, sizeof(double)); 
+	float *xReorder = (float* )calloc(localMatrixCOO.dimension, sizeof(float)); 
+	float *yReorder = (float* )calloc(localMatrixCOO.dimension, sizeof(float)); 
 	matrixReorder(&localMatrixCOO);
 	vectorReorder(localMatrixCOO.dimension, y, yReorder, localMatrixCOO.reorderList);
 
@@ -206,7 +206,7 @@ int main(int argc, char* argv[])
 
 	solverGPuUnprecondEHYB(&localMatrixCOO, yReorder, xReorder, MAXIter, &realIter);
 	
-	x = (double *) calloc(localMatrixCOO.dimension, sizeof(double));
+	x = (float *) calloc(localMatrixCOO.dimension, sizeof(float));
 	vectorRecover(localMatrixCOO.dimension, xReorder, x, localMatrixCOO.reorderList);
 
 	for (int i=0;i<10;i++)
@@ -229,7 +229,7 @@ int main(int argc, char* argv[])
 	//interval2=(end_time2-start_time2)*1000/CLOCKS_PER_SEC;
 
 	//printf("time consuming CPU is %f, time consuming GPU is %f, speedup is %f\n", interval1, interval2, interval1/interval2);
-	//double Gflop=(totalNum*4+12*dimension)/interval1*1000*MAXIter;
+	//float Gflop=(totalNum*4+12*dimension)/interval1*1000*MAXIter;
 	//printf("error is %f, total num is %d, time is %f ms, Gflops is %f, final error is %f\n",result_error/dimension, totalNum, interval1, Gflop, error_track[MAXIter-1]*1000);
 	return 0;
 }
